ISC license, contact, 3 space indent..
Massive adjustment for incompatible version 0.6.0 that is
in the queue: change from 3-clause BSD license to ISC, as
i am doing that for almost all my code at the moment and
i am not interested in the clauses as such anyway.
Change my mail address to the sourceforge one.
Use 3-space indent instead of 4-space.
Write «, i.e., » not «i.e.» in the POD part.