Change interface for convenience plus..
This commit introduces a massive change in the way S-SymObj
works, and breaks backward compatibility.

From now on S-SymObj will store the hash reference that will
be passed to sym_create(), so that the user does not need to
pass that again to obj_ctor().  The exclusion list has been
dropped too.

Instead the user must prefix to-be-excluded members with a
question mark, as in '?_member', when specifying the member
in the hash passed to sym_create(); the function will detect that
and modify the fields hash as necessary after not providing access
functions for these members (so that the member name follows the
regular _NAME scheme after the correction was applied).

Also apply a fix that yet caused constructor arguments to
become added to template member content instead of being
replacements for them.