Add "fast-constructor" paths, plus fixes..
Implement so-called "fast constructor" paths which bring a dramatical
performance improvement in case all classes of the hierarchy imple-
ment this optional feature (it automatically falls back to normal
operation otherwise).
It can do so because it creates one hash, fills in all necessary
fields from all the hierarchy in a single loop, and calls a
special function, which has to be provided via sym_create(), to
act as a "fast-constructor" for each class of the hierarchy.
Most often this can simply be an empty closure that does nothing.
This path also does not check args nor does any checks at all.

Unfortunately this commit also converts many block-local
temporary variables to function-local ones and change their
name to the more usual temporary names.

It also fixes (Add readonly accessors.., 2012-08-20), which
unfortunately missed to update the real reference in the
hash of the class, so the names yet included the "!"..