On our way to v0.6.0 -- call it a rewrite..
This changes the interface in that the user only calls
sym_create() and that does it all.  (Add "fast-constructor" paths,
plus fixes.., 2012-08-21) is obsolete in that new() is now
automatically created and switches dependent upon $Debug in
between a fast and slow, debug enabled version.  The user hook
remains in a more generic, user-constructor called version and is
now the only way to execute custom code in the constructor.

This change results in easier object creation, more debug checks
and a much, much faster non-$Debug execution path.  Also
per-object (-hierarchy) flags can now be used to debug a single
hierarchy only.

In short: S-SymObj < v0.6.0 was a somewhat convenient but crude
way to create objects during the debug phase.  S-SymObj v0.6.0 is
a rocking good way to create object hierarchies!