Add [@%] type modifiers..
which allow to use "undef" in the definition hash instead
of providing an empty array/hash just for the purpose
that S-SymObj gets the type of the field right and provides
the accessor sub.
The accessor subs themselves were already able to support
lazy allocation of the managed fields, but the actual logic
was yet missing.