Fix MI and deep tree hierarchies..
Hmm, originally SymObj was simply a file with two methods which were
copied over to the projects which needed them, so that the final
scripts were self-contained.
Once it was decided to make it a real perl(1) module and pimp it due
to that, it seems that some functionality regarding deep hierarchies
has been left off, for whatever reasons.

So do actually implement deep tree hierarchy support for SymObj.pm,
and fix some problems with overwrites through implicit references
(in 'foreach(@X){}', why is $_ implicitly $$_?  Here you go).