Make generic array/hash setters global subs..
Yet it was decided on a per-class level which generic setters
are required, and a per-class closure was then created as
necessary.

Drop these closures, use SymObj:: private global subs instead to
shrink the amount of instantiated code.  The cost of this
decision is of course that more arguments are needed to pass the
required information.